first load
templogger.py log a ds18b20 sensor temperature and send the result to a thingspeak channel
waterlogger.py log 2 float sensor (simple float sensor connected to 2 gpio ) and send 0 or 1 based on whather they are on or off. 

those scripts can be called by a cron every few minutes and we'll add smartness to them later.